requirements.txt
# 基础依赖
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2

# Web框架
Flask>=2.0.0
flask-cors>=3.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# 数据库
SQLAlchemy>=1.4.0
redis>=4.0.0

# 机器学习
torch>=1.9.0
statsmodels>=0.13.0

# 可视化
plotly>=5.0.0

# 工具
python-dotenv>=0.19.0
loguru>=0.5.3
click>=8.0.0
flasgger>=0.9.5

# 测试
pytest>=6.2.5

# 深度学习
torchvision>=0.10.0

# 时间序列分析
statsmodels>=0.12.2

# 数据可视化
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0

# 性能监控
prometheus-client>=0.11.0 